+++ app/models/tag.rb
@@ -1,3 +1,6 @@
 class Tag < ActiveRecord::Base
   belongs_to :document
+  
+  validates :text, presence: true
+  validates :document, presence: true
 end
